+++ index.html
@@ -28,7 +28,7 @@
     </div>
   </div>
 </body>
-<script src="js/script.js">
+<script type="module" src="js/index.js">
 </script>
 
 </html>
